Convert to use new get-file command; Convert the rest to use new vars; Configuration moved out; Script to clear and fetch; Much more...

+++ packages/linux-api-3.19/prepare.cmd.sh
@@ -1,4 +1,4 @@
-[ "$TARGET" = "initsys" ] || die "Package $NAME doesn't support $TARGET target!"
+[ "$TARGET" = "initsys" ] || [ "$TARGET" = "DEFAULT" ] || die "Package $NAME doesn't support $TARGET target!"
 
 rm -rf $SRC &> /dev/null
